requirements.txt
pytest==6.2.4
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
requests==2.25.1
PyYAML==5.4.1
autopep8==1.5.7
openpyxl==3.0.7
allure-pytest==2.9.43
jsonpath==0.82
pytest-sugar==0.9.4